[REFACTO] Typage strict global et correctifs de typage.

+++ sources/src/Controller/Admin/OrganizerCrudController.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Controller\Admin;
 
 use App\Entity\User;
@@ -103,9 +105,13 @@ class OrganizerCrudController extends AbstractController
         ]);
     }
 
-    #[Route('/delete/{id}', name: 'organizer_delete', methods: [Request::METHOD_GET, Request::METHOD_DELETE])]
+    #[Route('/delete/{id}', name: 'organizer_delete', methods: [Request::METHOD_POST, Request::METHOD_DELETE])]
     public function delete(User $user, Request $request): RedirectResponse
     {
+        if (! $this->isCsrfTokenValid('delete'.$user->getId(), (string) $request->request->get('_token'))) {
+            throw $this->createAccessDeniedException('Token CSRF invalide.');
+        }
+
         $this->em->remove($user);
         $this->em->flush();
